Submitting project "save state"
Great advances have been made. The page is almost
done with the exception of the form styling and
potentially some minor general styling details
after a final review.

The Odin logo has been added to the img folder,
which is used on the side image. A font folder
containing the fonts used on the page has been
added as well.